Data.Set.Ordered

Description

This module provides OSet, an insertion-order-preserving set, with type class instances for Foldable, Semigroup, Monoid and Data as well as a map function.

This is intended to be API-compatible with OSet in unordered-containers but with a few extra type class instances.

Documentation

(|>) infixl 5

Arguments

:: Ord a

=> OSet a
set

-> a
element

-> OSet a
set

\(O(log(N))\). Append an element to the end of set if the set does not already contain the element. The element is ignored if it is already in the set.

(|<) infixr 5

Arguments

:: Ord a

=> a
element

-> OSet a
set

-> OSet a
set

\(O(log(N))\) if the element is not in the set, \(O(N)\) if the element is already in the set. Prepend an element to the head of the set if the set does not already contain the element. The element is moved to the head of the sequence if the element is already present in the set.

data OSet a

An OSet behaves much like a Set but remembers the order in which the elements were originally inserted.

empty

Arguments

:: OSet a
set

\(O(1)\). The empty set.

filter

Arguments

:: (a -> Bool)
predicate

-> OSet a
set

-> OSet a
set

\(O(N)\). Filter a set by returning a set whose elements satisfy the predicate.

fromList

Arguments

:: Ord a

=> [a]
elements

-> OSet a
set

\(O(N log(N))\). Create a set from a finite list of elements. If an element occurs multiple times in the original list, only the first occurrence is retained in the resulting set. The function toList, \(O(N)\), in Foldable can be used to return a list of the elements in the original insert order with duplicates removed.

map :: Ord b => (a -> b) -> OSet a -> OSet b

\(O(N log(N))\). Return the set obtained by applying a function to each element of this set. Note that the resulting set may be smaller than the original. Along with the Ord constraint, this means that OSet cannot provide a lawful Functor instance.

member

Arguments

:: Ord a

=> a
element

-> OSet a
set

-> Bool
True if element is in set, False otherwise

\(O(log(N))\). Determine if the element is in the set.

notMember

Arguments

:: Ord a

=> a
element

-> OSet a
set

-> Bool
True if element is not in set, False otherwise

\(O(log(N))\). Determine if the element is not in the set.

singleton

Arguments

:: a
element

-> OSet a
set

\(O(1)\). A singleton set containing the given element.